1.7 (Build: 2015031001)
Added manual grade items that are added directly to the gradebook.
Doesn't show feedback or grades that have been hidden in the gradebook for an individual student.
Fixed Turnitin v2 assignments not displaying the type, submission or feedback links.
Added SQL to retrieve the gradetype (1=numeric) and scaleid in prepration for supporting letter and scale grades.
Known issue: some users have reported double-ups of data in both MySQL and Postgres, however the developer has been unable to replicate this problem so it is still outstanding.