2.3 (Build: 2015060803)
Responsive table layout and tooltips for mobile access - both phone and tablet. Uses footable (http://fooplugins.com/footable-demos/) to dymanically shrink the table to show only core information. The rest is displayed underneath with a drop down toggle to view it. Features expand all and collapse all toggles, search, pagination and course filter. Uses tooltips (http://osvaldas.info/elegant-css-and-jquery-tooltip-responsive-mobile-friendly) that display when hovered-over on desktops and when clicked-on on mobile devices. Moved some hard-coded text to the languages file to ensure cross-language compatibility.