2.7 (Build: 2016011900) bug fixes and new features including Personal Tutor tabs
Fixes
------
Fixed the issue further due to other issues popped up with some quiz submissions shown as late when they aren't
Fixed an issue with the self-reflective notes and turnitin feedback where an iteminstance had to be added to the DB table as turnitin uses the same grade item for multiple parts but used the subparts id for the different instances.
Fixed the decimal places to return the amount of decimal places as is set in Moodle.
Fixed the a further issue with max grade not updated when this is modified affter students are already graded.
Fixed the issue where tutors/admin could  edit or add to the 'related' user's self-reflective notes
Fixed the hardcoded link in the mytutees tab which was causing an error.
Fixed the mytutee page where users in inactive courses are showing in the tutee list and inactive courses showing in the course shortname list.
Fixed the display of self-reflective notes to only the students themselves, their personal tutor and program admin but not module tutor (unless you are also their personal tutor).
Fixed any possibility of Cross site scripting (XSS) when adding turnitin feedback and self-reflective notes.
Fixed the assignment and workshop feedback files which were showing an icon if the grade item has a feedbackfile for any user and not if the specific individual has a feedback file.
Fixed the decimal places to only show decimals if the grade has a fraction.
Fixed the issue where module tutors were seeing self-reflective notes which are only for the students personal tutors and programme admin.
Fixed turnitintool and turnitintooltwo feedback to take students and tutors to the individual feedback page.
Fixed the turnitin feedback to not show a link if no grade is yet added.

New Features
-------------
Added the ellipsis(...) styling for the long shortname, also for the mytutees course shortnames and ensure it doesn't wrap to two or more lines.
Added a background colour if the non-moolde feedback is student-added.
Added the user's name to the Feedback heading to say My feedback report for xxx.
Tutors and admin were now allowed to see their own my feedback report if they are a student on any course
Enlarged the size of the My students tab table so that relationship column in particular can show more module shortnames and also add a media query for smaller screens to default to the original size of max 420px.
Modified the quiz attempts to show the review page and for multiple attempts, show the feedback text with a link to show the last of the number of attempts.
Added a tooltip to the reflective notes header explaining to students that personal tutor and programme admin will be able to see their self-reflective notes.