2.8.4 (Build: 2016100400)
Added strings in code to the lang file.
New Departmental admin, Module tutor, Personal tutor dashboards.
New My students tab for staff to view personal tutees and students on courses.
Checks for extensions granted to individual students (or groups) so the Moodle assignment and quiz due date is accurate for each student.
Now displays an icon showing where rubrics and grading forms are being used in Moodle Assignments.
Allows for archived Moodle versions to be added to the report (warning: switching this on is NOT recommended, as there are some permission issues around using this).
Multiple bug fixes.
Thoroughly tested: a production quality release.