WR363070: Fix issues with the modal
The modal uses an older version of Bootstrap Modal than is bundled
with newer Moodles. There were issues with the modal on the newest
upstream branch as well. I've removed the bootstrap dependency and
updated the mark-up for Bootstrap 4. Now modal will open correctly,
the `show.bs.modal` event completes and the modal can be dismissed.