WR432962: Cast highestgrade to float
When highestgrade is fetched from the database, it is returned as a
string containing a float value. There is an `if
($record->highestgrade)` check before using this to calculate a grade
percentage, to avoid dividing by highestgrade if it is falsy (e.g "0").
However, if it is zero, the returned string is "0.00000", which is
truthy.

Casting the string to a float ensures the check performs as intended.