WR432962: Move event triggers earlier in the index page.
Lots of "Error writing to database" exceptions are being reported in the
shutdown function that writes event logs on this page. It's not clear
why, but it's possible that its because the event are being triggered
right at the end of the script, after the page output has been
completed.

Since the logs are buffered and written during the shutdown, there
should be no performance issue triggerring the events earlier, so
hopefully this is also prevent the database errors.